modified the service provider compoent UI

+++ service-your-desk-backend/src/main/java/com/service_your_desk/service_your_desk_backend/model/ServiceEntity.java
@@ -7,7 +7,7 @@ public class ServiceEntity {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer serviceId;
+    private Long serviceId;
     
     @Column(nullable = false, length = 100)
     private String name;
@@ -25,10 +25,10 @@ public class ServiceEntity {
         this.basePrice = basePrice;
     }
 
-    public Integer getServiceId() {
+    public Long getServiceId() {
         return serviceId;
     }
-    public void setServiceId(Integer serviceId) {
+    public void setServiceId(Long serviceId) {
         this.serviceId = serviceId;
     }
     public String getName() {
